requirements.txt
# Core framework
crewai>=0.28.0
crewai-tools>=0.1.0

# LLM providers
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0  # Pretty console output

# Optional: For web search tools
duckduckgo-search>=4.0.0

# Optional: For document processing (future RAG agent)
# chromadb>=0.4.0
# pypdf>=3.0.0
